requirements.txt
numpy
scipy
opencv-python
pillow
termcolor
pyyaml
tqdm
absl-py
trimesh
torchmetrics
seaborn
tensorboard
scikit-image
matplotlib
kornia
libigl
lpips
omegaconf
plyfile
seaborn
h5py
wandb
hydra-core

git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
submodules/simple-knn
submodules/r3dg-rasterization
submodules/radegs-rasterization
submodules/nvdiffrast
submodules/bvh
submodules/gs-ir
submodules/renderutils